requirements.txt
fastapi
uvicorn
python-dotenv
httpx
langchain
langchain-openai
langchain-community
langchain-text-splitters
langgraph
langgraph-checkpoint-postgres
google-generativeai
pypdf
openai
chromadb
psycopg[binary]
redis
kestra
bcrypt
